Moved model tests from controller folder to model folder
