+++ CVGS.Tests/CVGS.Tests.csproj
@@ -95,8 +95,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\EventModelTest.cs" />
-    <Compile Include="Controllers\GameModelTest.cs" />
+    <Compile Include="Models\EventModelTest.cs" />
+    <Compile Include="Models\GameModelTest.cs" />
     <Compile Include="Controllers\AccountControllerTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controllers\HomeControllerTest.cs" />
